tdlib: Rename action_list module to chat_action_list
